Restructure Folders for Consistency Between Platforms
