+++ bin/create
@@ -31,7 +31,7 @@
 
 const path = require('path');
 const ConfigParser = require('cordova-common').ConfigParser;
-const Api = require('./template/cordova/Api');
+const Api = require('./templates/cordova/Api');
 
 const argv = require('nopt')({
     help: Boolean,
@@ -52,7 +52,7 @@ if (argv.help || !projectPath) {
     console.log('   <project_template_dir>: Path to project template (override).');
     process.exit(0);
 } else {
-    const configPath = path.resolve(__dirname, 'template/config.xml');
+    const configPath = path.resolve(__dirname, 'templates/project/config.xml');
     const config = new ConfigParser(configPath);
 
     // apply overrides (package and project names
