+++ bin/templates/cordova/Api.js
