+++ bin/templates/cordova/build
