+++ bin/templates/cordova/build.bat
