+++ bin/templates/cordova/clean
