+++ bin/templates/cordova/clean.bat
