+++ bin/templates/cordova/defaults.xml
