+++ bin/templates/cordova/handler.js
