+++ bin/templates/cordova/lib/build.js
