+++ bin/templates/cordova/lib/check_reqs.js
