+++ bin/templates/cordova/lib/clean.js
