+++ bin/templates/cordova/lib/run.js
