+++ bin/templates/cordova/log
