+++ bin/templates/cordova/parser.js
