+++ bin/templates/cordova/run
