+++ bin/templates/cordova/run.bat
