+++ bin/templates/cordova/version
