+++ bin/templates/cordova/version.bat
