+++ bin/templates/project/config.xml
