+++ bin/templates/project/main.js
