+++ bin/templates/project/www/img/logo.png
