+++ bin/templates/project/www/img/splash.png
