+++ bin/templates/project/www/js/index.js
