+++ tests/spec/coverage.json
@@ -1,8 +1,9 @@
 {
-    "spec_dir": "spec",
+    "spec_dir": "tests/spec",
     "spec_files": [
         "unit/**/*[sS]pec.js"
     ],
     "stopSpecOnExpectationFailure": false,
     "random": false
 }
+
