+++ tests/spec/unit.json
@@ -1,0 +1,9 @@
+{
+    "spec_dir": "tests/spec",
+    "spec_files": [
+        "unit/**/*[sS]pec.js"
+    ],
+    "stopSpecOnExpectationFailure": false,
+    "random": false
+  }
+
